Added sig-storage labels to upgrade tests and moved them to appropriate directory
