+++ test/e2e/lifecycle/BUILD
@@ -29,6 +29,7 @@ go_library(
         "//test/e2e/framework/ginkgowrapper:go_default_library",
         "//test/e2e/upgrades:go_default_library",
         "//test/e2e/upgrades/apps:go_default_library",
+        "//test/e2e/upgrades/storage:go_default_library",
         "//test/utils:go_default_library",
         "//test/utils/junit:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
