+++ test/e2e/upgrades/BUILD
@@ -16,7 +16,6 @@ go_library(
         "horizontal_pod_autoscalers.go",
         "ingress.go",
         "mysql.go",
-        "persistent_volumes.go",
         "secrets.go",
         "services.go",
         "sysctl.go",
@@ -36,7 +35,6 @@ go_library(
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/util/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/uuid:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/wait:go_default_library",
     ],
@@ -54,6 +52,7 @@ filegroup(
     srcs = [
         ":package-srcs",
         "//test/e2e/upgrades/apps:all-srcs",
+        "//test/e2e/upgrades/storage:all-srcs",
     ],
     tags = ["automanaged"],
 )
